requirements.txt
# Core Dependencies
ccxt==4.2.25
pandas==2.2.3
numpy==1.23.5

# Technical Analysis
TA-Lib==0.6.8

# Data Validation and Settings
pydantic==2.8.2
pydantic-settings==2.4.0

# Scheduling and Monitoring
apscheduler==3.10.4
watchdog==5.0.2

# Database
redis==5.0.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Web Framework
fastapi==0.111.1
uvicorn==0.30.1
flask==3.0.3

# HTTP and Async
requests==2.32.3
aiohttp==3.10.5
websockets==13.1

# Configuration
python-dotenv==1.0.0

# Security
cryptography==42.0.8

# Built-in modules (no version needed)
asyncio

# Testing Framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-xdist==3.3.1
pytest-html==4.1.1
pytest-benchmark==4.0.0
